fix(decorators/localized-attr): set default value for new object api
We added a new api with wich you can access the translation object directly.
This deprecated the `getUnlocalizedField` method.

Now we set the default value to a empty object so on creation of a new model
you dont always have to assign an empty object yourself. This mostly fixes
reference lookup errors like ""[insertYourPropertyName].en was not found".